ref: Restructure C++ code into clear modules
Definition and implementation is separated into the following parts:

 - data_structures.h: Contains extern C headers for common types
 - data_definitions.h: Contains their mapping onto C++ types
 - processor.h: Specifies types and functions for minidump processing
 - resolver.h: Specifies types and functions for resolving

Note: resolver.h also contains delcarations for all data accessors
of resolved frames. Since the exported default processor only returns
unresolved frames, these accessors won't be needed before resolution.

Still, only a single library is built and always linked.
